data-owner-eth-client/requirements.txt
attrs==20.3.0
base58==2.1.0
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
importlib-metadata==3.7.0
importlib-resources==5.1.1
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.15.3
pycryptodome==3.10.1
pyrsistent==0.17.3
requests==2.25.1
rlp==2.0.1
six==1.15.0
toolz==0.11.1
typing-extensions==3.7.4.3
urllib3==1.26.3
varint==1.0.2
web3==5.17.0
websockets==8.1
zipp==3.4.0
kafka-python==2.0.1
requests==2.25.1